# config/schema/admin-openapi.yaml
openapi: 3.1.0
info:
  title: AegisProxy administration API
  version: 1.0.0
  description: >-
    Current v1 contract. This document is not served by the daemon. Unless a route explicitly
    says otherwise, the current implementation is available only through the private Unix socket.
    Typed Proxy Host validation and preview are non-persistent.
servers:
  - url: http://localhost
    description: HTTP over the configured private Unix socket
paths:
...
  /v1/access-policies:
    get:
      operationId: listAccessPolicies
      responses:
        "200":
          description: Stable object-ID ordered Access Policies owned by authenticated principal
          content:
            application/json:
              schema:
                type: array
                maxItems: 1024
                items: {$ref: "#/components/schemas/StoredAccessPolicy"}
        default: {$ref: "#/components/responses/Error"}
    post:
      operationId: createAccessPolicy
      parameters: [{$ref: "#/components/parameters/IfMatch"}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/AccessPolicyObject"}
      responses:
        "201":
          description: Owned Access Policy desired state persisted without runtime activation; object generation is returned as ETag
          headers:
            ETag:
              description: Quoted object generation
              schema: {type: string, pattern: '^"[1-9][0-9]*"$'}
          content:
            application/json:
              schema: {$ref: "#/components/schemas/StoredAccessPolicy"}
        default: {$ref: "#/components/responses/Error"}
  /v1/access-policies/{id}:
    get:
      operationId: getAccessPolicy
      parameters: [{$ref: "#/components/parameters/TypedObjectId"}]
      responses:
        "200":
          description: Owned Access Policy desired state and object-generation ETag
          headers:
            ETag:
              description: Quoted object generation
              schema: {type: string, pattern: '^"[1-9][0-9]*"$'}
          content:
            application/json:
              schema: {$ref: "#/components/schemas/StoredAccessPolicy"}
        default: {$ref: "#/components/responses/Error"}
    put:
      operationId: updateAccessPolicy
      parameters:
        - {$ref: "#/components/parameters/TypedObjectId"}
        - {$ref: "#/components/parameters/IfMatch"}
        - {$ref: "#/components/parameters/ObjectGeneration"}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/AccessPolicyObject"}
      responses:
        "200":
          description: Owned Access Policy desired state replaced without runtime activation
          headers:
            ETag:
              description: Quoted object generation
              schema: {type: string, pattern: '^"[1-9][0-9]*"$'}
          content:
            application/json:
              schema: {$ref: "#/components/schemas/StoredAccessPolicy"}
        default: {$ref: "#/components/responses/Error"}
    delete:
      operationId: deleteAccessPolicy
      parameters:
        - {$ref: "#/components/parameters/TypedObjectId"}
        - {$ref: "#/components/parameters/IfMatch"}
        - {$ref: "#/components/parameters/ObjectGeneration"}
      responses:
        "200":
          description: Owned Access Policy desired state deleted without runtime activation
          content:
            application/json:
              schema: {$ref: "#/components/schemas/StoredAccessPolicy"}
        default: {$ref: "#/components/responses/Error"}
...
  /v1/certificates:
    get:
      operationId: listCertificates
      responses:
        "200": {description: Owner-scoped typed Certificate objects}
        default: {$ref: "#/components/responses/Error"}
    post:
      operationId: createCertificate
      parameters: [{$ref: "#/components/parameters/IfMatch"}]
      requestBody: {$ref: "#/components/requestBodies/CertificateObject"}
      responses:
        "201": {description: Certificate object persisted}
        default: {$ref: "#/components/responses/Error"}
  /v1/certificates/{id}:
    get:
      operationId: getCertificate
      parameters: [{$ref: "#/components/parameters/TypedObjectId"}]
      responses:
        "200": {description: Owner-scoped typed Certificate object}
        default: {$ref: "#/components/responses/Error"}
    put:
      operationId: updateCertificate
      parameters:
        - {$ref: "#/components/parameters/TypedObjectId"}
        - {$ref: "#/components/parameters/IfMatch"}
        - {$ref: "#/components/parameters/ObjectGeneration"}
      requestBody: {$ref: "#/components/requestBodies/CertificateObject"}
      responses:
        "200": {description: Certificate object replaced}
        default: {$ref: "#/components/responses/Error"}
    delete:
      operationId: deleteCertificate
      parameters:
        - {$ref: "#/components/parameters/TypedObjectId"}
        - {$ref: "#/components/parameters/IfMatch"}
        - {$ref: "#/components/parameters/ObjectGeneration"}
      responses:
        "200": {description: Certificate object deleted}
        default: {$ref: "#/components/responses/Error"}
...
security:
  - bearerToken: []
components:
  securitySchemes:
    bearerToken:
      type: http
      scheme: bearer
      description: Optional on the private socket; local peer credentials remain mandatory.
    unixPeer:
      type: http
      scheme: aegis-unix-peer
      description: Operating-system peer credentials from the private Unix socket; no HTTP header is accepted.
    browserSession:
      type: apiKey
      in: cookie
      name: __Host-aegis-session
      description: Loopback-browser-only server-side session; bearer authentication is rejected.
  parameters:
    IfMatch:
      name: If-Match
      in: header
      required: true
      schema: {type: string, pattern: '^"[0-9]{20}-[0-9a-f]{64}"$'}
...
    TypedObjectId:
      name: id
      in: path
      required: true
      schema: {$ref: "#/components/schemas/ObjectId"}
    ObjectGeneration:
      name: X-Aegis-Object-Generation
      in: header
      required: true
      schema: {type: integer, minimum: 1}
...
  requestBodies:
...
    CertificateObject:
      required: true
      content:
        application/json:
          schema: {$ref: "#/components/schemas/CertificateObject"}
...
  responses:
    Error:
      description: Stable redacted error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
    ObjectId:
      type: string
      pattern: '^[a-z][a-z0-9_-]{0,63}$'
...
    MiddlewareRef:
      type: string
      pattern: '^[a-z][a-z0-9_-]{0,62}$'
...
    AccessPolicyObject:
      type: object
      additionalProperties: false
      required: [api_version, metadata, spec]
      properties:
        api_version: {const: v1}
        metadata:
          type: object
          additionalProperties: false
          required: [id, owner_id]
          properties:
            id: {$ref: "#/components/schemas/ObjectId"}
            owner_id: {$ref: "#/components/schemas/ObjectId"}
        spec:
          type: object
          additionalProperties: false
          required: [enabled, middlewares]
          properties:
            enabled: {type: boolean}
            shared_with:
              type: array
              maxItems: 128
              uniqueItems: true
              items: {$ref: "#/components/schemas/ObjectId"}
            middlewares:
              type: array
              minItems: 1
              maxItems: 64
              uniqueItems: true
              items: {$ref: "#/components/schemas/MiddlewareRef"}
    StoredAccessPolicy:
      type: object
      additionalProperties: false
      required: [generation, object]
      properties:
        generation: {type: integer, minimum: 1}
        object: {$ref: "#/components/schemas/AccessPolicyObject"}
    CertificateObject:
      type: object
      additionalProperties: false
      required: [api_version, metadata, spec]
      properties:
        api_version: {const: v1}
        metadata:
          type: object
          additionalProperties: false
          required: [id, owner_id]
          properties:
            id: {$ref: "#/components/schemas/ObjectId"}
            owner_id: {$ref: "#/components/schemas/ObjectId"}
        spec:
          type: object
          additionalProperties: false
          required: [enabled, certificate_ref]
          properties:
            enabled: {type: boolean}
            shared_with:
              type: array
              maxItems: 128
              uniqueItems: true
              items: {$ref: "#/components/schemas/ObjectId"}
            certificate_ref: {$ref: "#/components/schemas/ObjectId"}
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, details, request_id]
          properties:
            code: {type: string}
            message: {type: string}
            details: {type: array, items: {type: object}}
            request_id: {type: string}